+++ src/main/java/it/gov/pagopa/pu/workflow/wf/paymentsreporting/pagopa/wffetch/OrganizationPaymentsReportingPagoPaFetchWF.java
@@ -1,17 +1,17 @@
-package it.gov.pagopa.pu.workflow.wf.pagopa.paymentreporting.wffetch;
+package it.gov.pagopa.pu.workflow.wf.paymentsreporting.pagopa.wffetch;
 
 import io.temporal.workflow.WorkflowInterface;
 import io.temporal.workflow.WorkflowMethod;
 
 /**
- * Workflow interface for the Pago PA Payments Reporting Workflow
+ * Workflow interface for the Pago PA Payments Reporting Workflow per Organization
  *
  * */
 @WorkflowInterface
 public interface OrganizationPaymentsReportingPagoPaFetchWF {
 
   /**
-   * Workflow method for the Pago PA Payments Reporting Workflow
+   * Workflow method for the Pago PA Payments Reporting Workflow per Organization
    * @param organizationId the ID of the organization
    * */
   @WorkflowMethod
